Move session.*.php to Session class
